Explicitly cast size_t buffer size to int to avoid warning
The cast is safe because the buffer size is less than bufsize which is
much smaller than maximum size_t range, but doing it explicitly avoids
a clang-tidy warning.

+++ src/archive.cpp
@@ -555,7 +555,7 @@ bool Archive::Load(wxInputStream& stream)
     size_t len;
     while (status == XML_STATUS_OK &&
             (len = stream.Read(buf.data(), bufsize).LastRead()) != 0)
-        status = XML_Parse(parser, buf, len, len < bufsize);
+        status = XML_Parse(parser, buf, static_cast<int>(len), len < bufsize);
 
     XML_ParserFree(parser);
     return status == XML_STATUS_OK;
